Created spring JPA entity having many-to-one relationship, Created restful resource operations on JPA entity having many-to-one relationship.

+++ src/main/java/com/abhi/webservices/myrestwebservice/dao/VehicleDAO.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VehicleDAO extends JpaRepository<Vehicle, Long> {
+public interface VehicleDAO extends JpaRepository<Vehicle, Integer> {
 
 }
